LayoutTests/fast/events/ios/tab-cycle-expected.txt

This tests tab cycling through editable elements on iOS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS elementName is "A"
PASS elementName is "B"
PASS elementName is "C"
PASS successfullyParsed is true

TEST COMPLETE
